+++ concertseat/src/main/java/site/concertseat/domain/review/dto/req/ReviewListReq.java
@@ -1,4 +1,4 @@
-package site.concertseat.domain.review.dto;
+package site.concertseat.domain.review.dto.req;
 
 import lombok.Data;
