+++ concertseat/src/main/java/site/concertseat/domain/review/dto/res/ReviewListRes.java
@@ -1,7 +1,8 @@
-package site.concertseat.domain.review.dto;
+package site.concertseat.domain.review.dto.res;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import site.concertseat.domain.review.dto.ReviewDto;
 import site.concertseat.global.dto.SliceDto;
 
 @Data
